[BE] refactor: PlayingHistoryBuffer 인터페이스화 #752
