+++ backend/app/src/main/java/com/onair/hearit/app/playinghistory/infrastructure/buffer/PlayingHistoryBuffer.java
@@ -1,0 +1,10 @@
+package com.onair.hearit.app.playinghistory.infrastructure.buffer;
+
+import com.onair.hearit.core.domain.PlayingHistory;
+
+public interface PlayingHistoryBuffer {
+
+    void add(PlayingHistory playingHistory);
+
+    void flush();
+}
