+++ backend/app/src/main/java/com/onair/hearit/app/playinghistory/infrastructure/buffer/PlayingHistoryQueueBuffer.java
@@ -1,7 +1,7 @@
-package com.onair.hearit.app.playinghistory.infrastructure.scheduler;
+package com.onair.hearit.app.playinghistory.infrastructure.buffer;
 
-import com.onair.hearit.core.domain.PlayingHistory;
 import com.onair.hearit.app.exception.custom.BufferRequestException;
+import com.onair.hearit.core.domain.PlayingHistory;
 import com.onair.hearit.core.infrastructure.jdbc.PlayingHistoryCommandRepository;
 import java.util.ArrayList;
 import java.util.List;
@@ -14,19 +14,21 @@ import org.springframework.transaction.annotation.Transactional;
 
 @Component
 @RequiredArgsConstructor
-public class PlayingHistoryBuffer {
+public class PlayingHistoryQueueBuffer implements PlayingHistoryBuffer {
 
     private static final int BUFFER_SIZE = 10_000;
 
     private final PlayingHistoryCommandRepository playingHistoryCommandRepository;
     private final BlockingQueue<PlayingHistory> queue = new LinkedBlockingQueue<>(BUFFER_SIZE);
 
+    @Override
     public void add(PlayingHistory playingHistory) {
         if (!queue.offer(playingHistory)) {
             throw new BufferRequestException("큐가 가득 차서 요청을 처리할 수 없습니다.");
         }
     }
 
+    @Override
     @Transactional
     @Scheduled(fixedDelay = 1_000)
     public void flush() {
